requirements.txt
six
tensorflow-gpu
scipy
numpy
tqdm
smart-open
annoy
